Making the creation of routes optional
This is useful for me and probably to other people when feeding markdown data in pages that are already created. 
Would it make sense to include this as an option? 
I would argue so because this plugin in for markdown and not for let say blogging.
I can update the documentation accordingly of course :) Let me know.
Toon