Added Base64 Image ID to help with Lighthouse.
Lighthouse sometimes dislikes ElderJS's base64 placeholder pictures.
Usually, it lets you know that the placeholder needs to be resized.
Unfortunately, Lighthouse doesn't help you identify the image well.
Which means on pages with lots of pictures, this is a pain.
Having an ID embedded in the data URL fixes this.
It does make pages slightly larger, but I feel the tradeoff is worth it.